git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-1.4@1516 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/residfp-builder/residfp/Filter6581.cpp
@@ -36,7 +36,7 @@ Filter6581::~Filter6581()
 
 void Filter6581::updatedCenterFrequency()
 {
-    const unsigned int Vw = f0_dac[fc];
+    const unsigned short Vw = f0_dac[fc];
     hpIntegrator->setVw(Vw);
     bpIntegrator->setVw(Vw);
 }
